Added calculation of separation fidelity
Added methods for finding the measurement separation fidelity along a readout quadrature. Fits the data to a double Gaussian distribution and finds the intersection point between the two Gaussians to define a separation axis. Error is calculated as 1-e_g-e_e the error caused by the overlap of the distributions.